Refactor image processing into its own component
